Rationalize usage of integers
  - Except at a few places, int have been replaced by dynamic_graph::size_type,
    also known as Eigen::Matrix::Index.
  - unsigned int have been replaced by std::size_t.

+++ src/tools/time-stamp.cpp
@@ -84,7 +84,7 @@ double &TimeStamp::getTimeStampDouble(const dynamicgraph::Vector &vect,
   sotDEBUGIN(15);
 
   if (offsetSet)
-    res = (vect(0) - offsetValue) * 1000;
+    res = (vect(0) - (double)offsetValue) * 1000;
   else
     res = vect(0) * 1000;
   res += vect(1) / 1000;
